Restructure program files
